graph: Move web3 type conversions into their own module
